+++ packages/core/src/utils/parseCEL/parseCEL.test.ts
@@ -596,9 +596,9 @@ it('handles custom expressions', () => {
     'opted_in_at.isBirthday(-1)',
     wrapRule({
       field: 'opted_in_at',
-      operator: 'isBirthday',
+      operator: 'isBirthday' as DefaultOperatorName,
       value: -1,
-    } as unknown as DefaultRuleType),
+    }),
     {
       customExpressionHandler: expr => {
         if (